requirements.txt
portia-sdk-python==0.4.2
python-dotenv==1.1.0
